move database functions to utils folder
